fix(#690): move rpc_decorator and context_utils from core/ to lib/
Both modules are pure utilities with zero kernel dependencies — they
belong in the tier-neutral nexus.lib package, not in core/.

- rpc_decorator: @rpc_expose decorator (zero deps, used by 18 services)
- context_utils: get_zone_id/get_user_identity/get_database_url helpers
- Also fixes 5 pre-existing type: ignore[attr-defined] in rpc_decorator

All 28 import sites updated. Old core/ files deleted (no backward compat).
