+++ src/nexus/lib/context_utils.py
@@ -1,13 +1,11 @@
-"""
-Utility functions for extracting and resolving context information.
+"""Utility functions for extracting and resolving context information.
 
-This module provides centralized helpers for:
+Tier-neutral utility (``nexus.lib``) — zero kernel dependency.
+
+Provides centralized helpers for:
 - Extracting zone_id from context with defaults
 - Extracting user identity (type, id) from context
 - Resolving database URLs with environment variable priority
-
-These utilities eliminate code duplication across nexus_fs_mounts, nexus_fs_oauth,
-and workspace_registry modules.
 """
 
 import logging
