+++ src/nexus/services/mcp_service.py
@@ -15,7 +15,7 @@ import builtins
 import logging
 from typing import TYPE_CHECKING, Any
 
-from nexus.core.rpc_decorator import rpc_expose
+from nexus.lib.rpc_decorator import rpc_expose
 from nexus.services.protocols.filesystem import NexusFilesystem
 
 logger = logging.getLogger(__name__)
