+++ src/nexus/services/memory_service.py
@@ -18,7 +18,7 @@ import logging
 from typing import TYPE_CHECKING, Any, Literal
 
 from nexus.contracts.types import OperationContext
-from nexus.core.rpc_decorator import rpc_expose
+from nexus.lib.rpc_decorator import rpc_expose
 
 if TYPE_CHECKING:
     from collections.abc import Callable
