+++ src/nexus/services/rebac_service.py
@@ -24,7 +24,7 @@ from datetime import datetime
 from typing import TYPE_CHECKING, Any, TypeVar
 
 from nexus.contracts.exceptions import CircuitOpenError
-from nexus.core.rpc_decorator import rpc_expose
+from nexus.lib.rpc_decorator import rpc_expose
 from nexus.services.rebac_share_mixin import ReBACShareMixin
 
 logger = logging.getLogger(__name__)
