+++ src/nexus/services/search_listing_mixin.py
@@ -22,7 +22,7 @@ from typing import TYPE_CHECKING, Any
 from nexus.constants import ROOT_ZONE_ID
 from nexus.contracts.exceptions import PermissionDeniedError
 from nexus.contracts.types import Permission
-from nexus.core.rpc_decorator import rpc_expose
+from nexus.lib.rpc_decorator import rpc_expose
 
 # Constants duplicated from search_service to avoid circular import
 LIST_PARALLEL_MAX_DEPTH = 100
