+++ tests/unit/server/test_rpc_admin_only.py
@@ -12,7 +12,7 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from nexus.core.rpc_decorator import rpc_expose
+from nexus.lib.rpc_decorator import rpc_expose
 
 # ============================================================================
 # Test 1: Decorator sets _rpc_admin_only flag
