+++ tests/unit/server/test_rpc_parity.py
@@ -378,7 +378,7 @@ def test_all_public_methods_are_exposed_or_excluded():
                 "",
                 "1. Add @rpc_expose decorator to the method (RECOMMENDED):",
                 "   ```python",
-                "   from nexus.core.rpc_decorator import rpc_expose",
+                "   from nexus.lib.rpc_decorator import rpc_expose",
                 "",
                 "   @rpc_expose(description='Your description')",
                 "   def your_method(self, ...):",
